polish(chat): rename tool-part-parsing to message-part-parsing, tidy exports
Zero behavior change. Renames tool-part-parsing.ts to
message-part-parsing.ts since it owns text/reasoning/file predicates
alongside tool parts, not just tool parts; drops the unused
ReasoningPartLike/TextPartLike re-export from conversation.ts; documents
why the ChatProviderModelInputMessage import in tool-replay-reconciliation.ts
must stay type-only; and inlines isTransientToolState as a direct export
instead of a trailing export block. Re-pins the FILE_SIZE_CEILINGS in
ban-chat-antipatterns.ts to match.
